requirements.txt
django==4.1.3
django-registration==3.3
djangorestframework==3.14
djoser==2.1.0